+++ resources/views/pages/partials/blocks/image.blade.php
@@ -1,21 +1,37 @@
 @php
-    $imageSource = $block->asset?->url();
+    $image = $block->media;
+    $imageSource = $image?->url();
+    $caption = trim((string) ($block->title ?? ''));
+    $altText = trim((string) ($block->subtitle ?? ''));
+    $fallbackAltText = trim((string) ($image?->alt_text ?: $image?->title ?: $caption ?: 'Image'));
+    $resolvedAltText = $altText !== '' ? $altText : $fallbackAltText;
+    $href = trim((string) ($block->url ?? ''));
+    $linkAttributes = '';
+
+    if ($href !== '' && preg_match('/^(https?:\/\/|\/|#|mailto:|tel:)/i', $href)) {
+        $linkAttributes = ' href="'.e($href).'"';
+    } else {
+        $href = '';
+    }
 @endphp
 
 @if ($imageSource)
-    <figure class="wb-stack wb-gap-2">
-        <img src="{{ $imageSource }}" alt="{{ $block->asset?->alt_text ?: $block->subtitle ?: $block->title ?: 'Image block' }}">
+    <figure class="wb-stack wb-gap-2" data-wb-public-block-type="{{ $block->publicBlockTypeAttribute() }}">
+        @if ($href !== '')
+            <a{!! $linkAttributes !!}>
+        @endif
+        <img
+            src="{{ $imageSource }}"
+            alt="{{ $resolvedAltText }}"
+            @if ($image?->width) width="{{ $image->width }}" @endif
+            @if ($image?->height) height="{{ $image->height }}" @endif
+        >
+        @if ($href !== '')
+            </a>
+        @endif
 
-        @if ($block->title)
-            <figcaption>{{ $block->title }}</figcaption>
+        @if ($caption !== '')
+            <figcaption>{{ $caption }}</figcaption>
         @endif
     </figure>
-@endif
-
-@if ($block->children->isNotEmpty())
-    <div class="wb-stack wb-gap-4">
-        @foreach ($block->children as $child)
-            @include('pages.partials.block', ['block' => $child])
-        @endforeach
-    </div>
 @endif